refactor: separate as utils/useVideoResourceProps.ts
